Move test in tests directory
