+++ packages/search-metadata-previews/tests/SnippetEditorTest.js
@@ -1,18 +1,14 @@
-import SnippetEditor from "../components/SnippetEditor";
+import SnippetEditor from "../src/SnippetEditor/components/SnippetEditor";
 import React from "react";
 import { createComponentWithIntl } from "@yoast/components";
 import {
 	mountWithIntl,
 	shallowWithIntl,
 } from "yoast-components/utils/helpers/intl-enzyme-test-helper";
-import {
-	MODE_DESKTOP,
-} from "../../SnippetPreview/constants";
-import {
-	focus,
-} from "../components/ReplacementVariableEditorStandalone";
+import { MODE_DESKTOP } from "../src/SnippetPreview/constants";
+import { focus } from "../src/SnippetEditor/components/ReplacementVariableEditorStandalone";
 
-jest.mock( "../components/ReplacementVariableEditorStandalone" );
+jest.mock( "../src/SnippetEditor/components/ReplacementVariableEditorStandalone" );
 
 const defaultData = {
 	title: "Test title",
