+++ packages/search-metadata-previews/tests/SnippetPreviewTest.js
@@ -1,5 +1,5 @@
-import SnippetPreview from "../components/SnippetPreview";
-import { MODE_DESKTOP, MODE_MOBILE } from "../constants";
+import SnippetPreview from "../src/SnippetPreview/components/SnippetPreview";
+import { MODE_DESKTOP, MODE_MOBILE } from "../src/SnippetPreview/constants";
 import React from "react";
 import { createComponentWithIntl } from "@yoast/components";
 import {
